cut down routing.Location to make room for more near relays

+++ modules/ghost_army/types.go
@@ -291,8 +291,8 @@ func (self *Entry) Into(data *transport.SessionPortalData, dcmap DatacenterMap, 
 		meta.DeltaRTT = deltaRTT
 
 		meta.Location = routing.Location{
-			Latitude:  self.Latitude,
-			Longitude: self.Longitude,
+			Latitude:  float32(self.Latitude),
+			Longitude: float32(self.Longitude),
 			ISP:       self.ISP,
 		}
 
@@ -345,7 +345,7 @@ func (self *Entry) Into(data *transport.SessionPortalData, dcmap DatacenterMap, 
 	// map point
 	{
 		pt := &data.Point
-		pt.Latitude = self.Latitude
-		pt.Longitude = self.Longitude
+		pt.Latitude = float32(self.Latitude)
+		pt.Longitude = float32(self.Longitude)
 	}
 }
